NCBIAccession collapse + vroom patch
Removed the step from the PREDICT pipeline that collapses the accessions, and then added it at the very end of the pipeline for unique-otherwise records. Also added col_type specification for two very sparse columns that seem to get read as logical instead of double by vroom, leading to the loss of PMID and PublicationYear in the final copy. This is something users also need to be wary of when they read the data in.